fixed #8
